+++ modules-services2/service-templates-core/src/main/java/org/openprovenance/prov/service/core/Querier.java
@@ -51,7 +51,7 @@ public class Querier  {
         StringBuilder sb=new StringBuilder();
         composeQuery.accept(sb,data);
         String statement=sb.toString();
-        logger.info("Executing statement: " + statement);
+        logger.debug("Executing statement: " + statement);
         boolean rs;
         try {
             rs=storage.executeStatements(conn,statement);